Feature ETP-1059: Evaluate expr in metadata
