+++ packages/MainUI/app/App.tsx
@@ -4,11 +4,8 @@ import SanityChecker from '../components/SanityChecker';
 import LanguageProvider from '../contexts/languageProvider';
 import MetadataProvider from '../contexts/metadata';
 import UserProvider from '../contexts/user';
-import useOBUtilities from '@/hooks/useOBUtilities';
 
 export default function App({ children }: React.PropsWithChildren) {
-  useOBUtilities();
-
   return (
     <LanguageProvider>
       <SanityChecker>
