--- packages/MainUI/hooks/useOBUtilities.ts
@@ -1,14 +1,0 @@
-import { useEffect } from 'react';
-
-export default function useOBUtilities() {
-  useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (window as any).OB = {
-      Utilities: {
-        getValue: (object: Record<string, never>, property: string) => {
-          return object[property];
-        },
-      },
-    };
-  }, []);
-}
